update example scripts for minimal text request with user input without dependencies
